\name{solarPos-package}
\alias{solarPos-package}
\docType{package}
\title{Solar Position Algorithm for Solar Radiation Applications}
\description{
Implementation of the Solar Position Algorithm for Solar Radiation Applications of Reda and Andreas (2004, 2007) for calculation of solar zenith and azimuth angles. 
}
\details{
\tabular{ll}{
Package: \tab solarPos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-07\cr
License: \tab GPL-2\cr
}
}
\author{
Jasper Van doninck

Maintainer: Jasper Van doninck <vandoninck.jasper@gmail.com>
}
\references{
Reda, I. and Andreas, A. (2004) Solar position algorithm for solar radiation applications, \emph{Solar Energy} 76, 577-589.

Reda, I. and Andreas, A. (2007) Corrigendum to Solar position algorithm for solar radiation applications [Solar Energy 76 (2004) 577589], \emph{Solar Energy} 81, 838.
}
